New makefile approach for RivieraLpmService. (#14)
Instead of Eddie cd'ing into components/RivieraLpmService to run
cmake as its own project, Eddie's CMakeLists.txt uses
ADD_DIRECTORY to reference RivieraLpmService.

This new approach fixes `components set-location` and as a bonus
simplifies the makefiles.  Also gives the product more
flexibility to determine how the LPM source files are compiled.

There are corresponding changes in RivieraLpmService.git.

Note that the BOSE_WORKSPACE environment variable is understood
by the CastleTools `components` script.  That's an important
detail for this to work.