Move BoseLogs into /data.
/mnt/nv/BoseLogs is now a symlink to /data/logs.

In sm2/scm, /mnt/nv was a 32MB compressed file system.  In
Riviera, /mnt/nv is still 32MB but it's not compressed, and so
two core files are enough to fill it.  /data is another file
system that the Riviera-HSP team recommended we could use.  But
moving log snapshots and core files to /data, we make it less
likely we'll fill up /mnt/nv and cause the system to misbehave.