Compress core files.
In sm2/scm the /mnt/nv file system was compressed.  In Riviera
it's not, and so we need to compress core files before writing to
disk, otherwise two core files can easily fill the whole 32MB
file system.

core files seem to compress by about 99%.  The AVS daemon for
example creates a 161KB uncompressed core file which compresses
to 2.4KB.