Handle symbol stripping differently.
All bin/* and lib/* files in the product.ipk file are now
stripped.  We save the unstripped binaries in
product-unstripped.tar.gz for later debugging (deployed by EC as
eddie_unstripped_<VERSION_STRING>.tar.gz).

Other components should store the unstripped executables and
libraries in their deployed tarballs.  If not, create-product-ipk
emits a warning like this:

    create-product-ipk: (warning) No debug symbols: .../shepherdd

There are several such warnings for now which we'll want to fix
by modifying each of the components.