PtsServer: Child handler may hang
In SHELBY-45616 there's a report that certain perturbations of
the network interfaces can cause the PtsServer to hang.

This changeset makes two changes to the operation of PtsServer:

1. When the PtsServer daemon launches a child to handle an HTTP
request, don't wait for the child to finish.

2. Put an upper limit on the amount of time a child can take to
handle a given HTTP request (60 seconds).

So, if a child hangs, the PtsServer will still be able to handle
other requests, and the timeout (alarm) will ensure that the
child eventually gets cleaned up.

60 seconds is presumed to be much longer than we should ever need
to handle a given HTTP request.

Also, PtsServer was ignoring SIGPIPE but there's no reason to be
doing that. The main daemon process never writes to
sockets/pipes. And if the child process were ever to write to a
closed socket, we'd want SIGPIPE to end the child. So, SIGPIPE is
no longer ignored.

http://reviewboard.bose.com/r/14666/