Enable LPM tap always. (#256)
Ordinarily, if the mfgdata is missing or invalid, we don't start
any of the SoundTouch daemons.  The idea is we want to make it
clear that the unit is configured wrong.

But we still want to be able to access the LPM tap so that we can
recover the unit.  With this commit, if the mfgdata is
missing/invalid, we still start the LPMService and CLIServer (but
nothing else).

This commit also moves the validation of the mfgdata out of the
SoundTouch script into its own script validate-mfgdata.

CASTLE-7265