Rework "LPM CLI only" mode.
If the manufacturing data is invalid, we don't launch all the
SoundTouch daemons, instead we launch only enough daemons to
enable the LPM CLI.  In this commit that's still true.  But now
with this commit, Shepherd launches those daemons.  This change
lets us do `stop;start` after correcting the mfgdata.