First working version.
Note CustomProductControllerStateBooting::HandleCapsState() doesn't call SendInitialCapsData() like the Common class does, so there is no effect. This will change with the expected rework of the state machine to make use of common state code.
The code here was tested with temporary code to call SendInitialCapsData() and test.html showed the expected values in /system/sources.