Add Test Rail and Device ID
New feature that allows individual pytest test cases to be linked
directly to a TestRail test case. A user should be able to just pip
install the requirements.txt file and continue on without issue. In
order to properly push to TestRail, more test configuration details need
to be added.

Add command line argument to specify the Android Device ID.