Start adbd if development=true
In an upcoming Riviera-HSP release, adbd will no longer be
launched automatically by the Riviera HSP. With this commit, the
SoundTouch script will launch adbd if development mode is
enabled. (If adbd is already running for whatever reason, trying
to start it again has no effect except to emit some error
messages.)

With this commit, we also kill adbd if development=false.  After
the new HSP is available, this step can be deleted.