PGC-2135 fix
This is an update to LPM to pull in the BLERemoteXCVR change that fixes PGC-2135  "Professor bond to remote fails after linking attempts".  This was the result of something called a Message Integrity Check (MIC) failure during re-pairing after an out of range condition or possibly a battery replacement scenario.