Compress the encrypted /pts and /logread output. (#1008)
* Compress the encrypted /pts and /logread output.

The downloaded file is now compressed using gzip and given a .dat
file extension.  The idea is to coax browsers into saving
directly to disk instead of displaying to the user (after which
the user might mangle it thereby making it useless).

Tested with Chrome, Internet Explorer and Edge.  Chrome stored it
straight to disk.  IE and Edge both prompted to ask what to do
and [save] was an option.

I've also modified the CastleTools/bin/decrypt-logs to detect if
the file is compressed and to uncompress it using gunzip before
decrypting.
https://github.com/BoseCorp/CastleTools/commit/1f3767dd9f2318652973d9107477b0c08205ee65

CASTLE-19514

* pts: Fix IE/Edge warning on /dev page.