Include pts-info at the end of pts-logs.
Might as well since it's a small amount of output compared to the
logread output and then users/customers have only one file to
download to get all of the /pts.dat and /logread.dat info.

This makes the PTS server endpoint /pts.dat redundant.

Per discussion with Jason Furgal.