validate-version: Commit IDs are >=7 chars. (#1111)
* validate-version: Commit IDs are >=7 chars.

An abbreviated commit ID is always at least 7 chars.  But if git
determines that 7 chars would be ambiguous, git uses 8 or more
chars for the abbreviated ID.

* Made some more refinements to the regex.

In `x{1}` is the same as just `x`.

`x{0,1}` is the same as `x?`.

Commit IDs are always lowercase hex digits, so `0-9a-zA-Z` should
be `0-9a-f`.

Build numbers will likely one day be more than 4 digits.
No need to limit to 4.