Don't kill/start BT on suspend/resume
"In the current drop of BT code that we have it should work
without killing the processes across suspend/resume. Can you make
the appropriate changes in the shepherd processing to not kill
and restart the BT processes and work with Matt G on a plan to
test this." -- Jon Cooper.