ProductConfig.json: New.
The design in other products is to use ProductConfig.json to
configure whether certains features are enabled or not based on
the product variant as detected at runtime.

For some reason, this design hasn't found its way to
professor/master yet.

Components are starting to expect to be able to find the file
`/opt/Bose/opt-bose-fs/etc/ProductConfig.json`.

For more context, see
https://github.com/BoseCorp/AudioSource-Spotify/pull/44