Change BoseWebMonaco back to 5.0.2
BoseWebMonaco was updated to the latest 5.1.0 build during an automatic update of all component.json dependencies. 5.1.0 was supposed to go in with LightbarController changes to support animation arrays. A decision was made to hold off on that change but the build still exists. For now we need to disregard 5.1.0 and and instead point to 5.0.2