Integrate telemetry into product controller.
The main change of this commit is to add the infrastructure so
the product controller can start collecting telemetry data.

In this commit, there are four new counters: network standby
enter & leave, low power standby & resume.

The intention is that these changes will get merged from Eddie to
the other products and we can add more counters, perhaps some
that are product specific (e.g., battery info for Taylor).

CASTLE-34814