Professor: Not saving debug symbols
Save the unstripped binaries for later debugging.

This apparently got broken on May 24 when we switch to using the
strip-symbols script from CastleTools.

Also sync'ed some changes in collect-deployables from
eddie/master.  But the relevant change in collect-deployables is
to save all `*-unstripped.tar.gz` files:

./builds/deploy/Alpha/product-unstripped.tar.gz
./builds/deploy/Alpha/IoTService-unstripped.tar.gz
./builds/deploy/Alpha/AVSService-unstripped.tar.gz

But the last two will require an upgrade to the IoTService and
AVSService components that isn't yet on their stable branches.

CASTLE-35168