Allow core files to be created in docker
I added two priv_* scripts to start docker with the --privileged
option and setup for core dumps. To get core dumps in a docker
container, you need to modify the /proc/sys/kernel/core_pattern.
To be able to modify files in /proc, you need to have special
priv's in the docker container.